ingest/pyproject.toml
[project]
name = "ingest"
version = "0.1.0"
description = "Document ingestion pipeline for beach jurisprudence RAG"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34",
    "docling>=2.0",
    "langchain>=0.3",
    "langchain-google-genai>=3.0",
    "langchain-docling>=2.0",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[tool.uv]
package = false
